added client-server connection setup and usernames
I know Daniel made two new files to integrate the networking part with the game part, but I decided it would to easier to not redo the game part and instead just start squishing in the networking parts into board-game-cli.py.

I also added the ability for players to enter a username. They don't have a username immediately after connecting to the server, so the automatically-generated id is used to identify them. In the code we add in the future, they could also be uniquely identified by their username, assuming we don't allow duplicate usernames.