file structure overhaul + multiple ctk screens
asdl[fjhpoh[wiehf[iausdofpi;a oijsdklfj;alksdj (12)

Let's all only add code to these files for the sake of organization and sanity. I think this is a decent file structure, so let's really try to stick to it. If you hack together some code in a new file, just keep it on your local machine.